Line alignment on block composition (REQ 3455424).
git-svn-id: http://svn.code.sf.net/p/clown/code/trunk@71 68affd6e-c39b-4d6d-bb22-b3492c778056

+++ dotNET/pdfclown.samples.cli/src/org/pdfclown/samples/cli/ParsingSample.cs
@@ -85,7 +85,8 @@ namespace org.pdfclown.samples.cli
       Pages pages = document.Pages;
       int pageCount = pages.Count;
       Console.WriteLine("\nPage count: " + pageCount);
-      int pageIndex = (int)Math.Round((float)pageCount / 2);
+
+      int pageIndex = (int)Math.Round(pageCount / 2d);
       Console.WriteLine("Mid page:");
       PrintPageInfo(pages[pageIndex],pageIndex);
